libs/dagster-io/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dagster-io"
version = "0.2.0"
description = "Shared IO manager, LLM resources, and utilities for Dagster pipelines"
requires-python = ">=3.12"
dependencies = [
    "dagster>=1.9",
    "boto3>=1.34",
    "pydantic>=2.0",
    "langchain-core>=0.3",
    "langchain-openai>=0.3",
    "langchain-text-splitters>=0.3",
    "pyyaml>=6.0",
    "prometheus-client>=0.20",
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
    "opentelemetry-exporter-otlp-proto-grpc>=1.20",
]

[project.optional-dependencies]
huggingface = [
    "langchain-huggingface>=0.1",
    "sentence-transformers>=3.0",
]
dev = [
    "pytest>=8.0",
    "ruff>=0.8",
]

[tool.hatch.build.targets.wheel]
packages = ["src/dagster_io"]

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
